==> vignettes/4_therapeutic_protocol.Rmd <==
---
title: "Therapeutic Protocol Definition"
author: "Niklas Rindtorff"
date: "3/14/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(tidyverse)
library(stringr)
library(here)
```

# Import data

```{r}
load(here("data/crxg.Rdata"))
load(here("data/crxg_map_imputed.Rdata"))
load(here("data/mutation.Rdata"))
load(here("data/cnv.Rdata"))
load(here("data/expression.Rdata"))
```

# Defining selection criteria for treatment arms

I start out allocating patients that will receive Palbociclib. 

```{r}
#palbociclib is more complicated - patients need a positive Rb expression and amplification of CCND1, CCND2 or CCND3
# an alternative inclusion criterium is CDK4 or CDK6 amplification

palbociclib_cnv <- cnv %>% 
  dplyr::select(contains("ccnd"), contains("cdk4"), contains("cdk6"), cosmic_id) %>% 
  mutate(palbociclib_cnv = gain_cna_pancan59_ccnd1_cttn + gain_cna_pancan165_cdk4,
         cosmic_id = as.character(cosmic_id))
  
rb_expression <- expression %>% 
  ungroup() %>%
  dplyr::select(ENSG00000139687, cosmic_id) %>% 
  mutate(ENSG00000139687 = as.character(ENSG00000139687) %>% as.numeric()) %>% 
  as_tibble()

cutoff_rb <- quantile(rb_expression$ENSG00000139687)[2]

rb_expression %>% 
  ggplot(aes(ENSG00000139687)) + 
  geom_density() + 
  theme_classic() + 
  geom_vline(xintercept = cutoff_rb)

palbociclib_inclusion <- rb_expression %>% 
  mutate(palbociclib_exp = if_else(ENSG00000139687 > cutoff_rb, 1, 0)) %>%
  left_join(palbociclib_cnv)
           

```

I continue allocating patients that will receive a targeted BRAF V600E inhibitor, *Dabrafenib*. 

```{r, message = FALSE, warning=FALSE, eval = FALSE}
library(readxl)
wes_variants <- read_excel(here("local_data/WES_variants.xlsx"), 
    sheet = "WES_variants")

save(wes_variants, file = here("data/wes_variants.Rdata"))
```

```{r}
load(here("data/wes_variants.Rdata"))

v600_data <- wes_variants %>% 
  #filter(Gene == "BRAF") %>% 
  mutate(v600 = if_else(grepl(AA, pattern = "p.V600") & Gene == "BRAF", 1, 0),
         COSMIC_ID = as.character(COSMIC_ID)) %>% 
  dplyr::select(cosmic_id = COSMIC_ID, v600) %>% 
  nest(-cosmic_id) %>% 
  mutate(new = purrr::map(data, ~ .x %>% arrange(desc(v600)) %>% head(1))) %>% 
  unnest(new) %>% 
  dplyr::select(-data) %>%
  distinct()
```


I complete my allocation.

```{r}
allocation <- mutation %>% 
  left_join(palbociclib_inclusion, by = "cosmic_id") %>%
  left_join(v600_data, by = "cosmic_id") %>%
  #dplyr::select(EGFR, ERBB2, PTCH1) %>%
  mutate(strict_arm = case_when(EGFR == 1 ~ "afatinib",
                         ERBB2 == 1 ~ "afatinib",
                         palbociclib_exp == 1 & palbociclib_cnv == 1 ~ "palbociclib",
                         PTCH1 == 1 ~ "vismodegib",
                         BRAF == 1 & v600 == 1~ "dabrafenib",
                         BRAF == 1 & v600 == 0 ~ "trametinib", # I need to filter BRAFV600 mutants
                         NF1 == 1 ~ "trametinib",
                         GNA11 == 1 ~ "trametinib",
                         BRCA1 == 1 ~ "olaparib",
                         BRCA2 == 1 ~ "olaparib",
                         TRUE ~ "cisplatin"),
         cosmic_id = as.numeric(cosmic_id)) %>% 
  dplyr::select(cosmic_id, strict_arm) %>% 
  distinct()

```


I recreate the heatmap from before, but include the assignment mechanism. 

```{r}
types_of_interest <- (crxg %>% 
    drop_na() %>%
    dplyr::select(cosmic_id, cancer_type ) %>% 
    distinct() %>% count(cancer_type) %>% arrange(desc(n)) %>% filter(n > 40) %>% .$cancer_type)


anno_col  = crxg %>% drop_na() %>%
                  dplyr::select(cosmic_id, cancer_type) %>% 
                  distinct() %>%
  mutate(cancer_type = if_else(cancer_type %in% types_of_interest,
    cancer_type, "Other")) %>%
  left_join(allocation) %>%
  mutate(strict_arm = ifelse(is.na(strict_arm), "not_identified", strict_arm)) %>%
  as.data.frame() %>%
  remove_rownames() %>%
 # dplyr::select(-cancer_type) %>%
  column_to_rownames("cosmic_id")

anno_row  = crxg %>% 
  drop_na() %>%
                  dplyr::select(drug_name, target_pathway) %>% 
                  distinct() %>%
  as.data.frame() %>%
  remove_rownames() %>%
                  column_to_rownames("drug_name")

# defining colors - always fun
cancer_type_color <- RColorBrewer::brewer.pal(length(types_of_interest)+1, "Paired")
names(cancer_type_color) <- c(types_of_interest, "Other")

target_pathway_color <- RColorBrewer::brewer.pal(length(anno_row$target_pathway %>% unique()), "Accent")
names(target_pathway_color) <- anno_row$target_pathway %>% unique()

strict_arm_color <- RColorBrewer::brewer.pal(length(anno_col$strict_arm %>% unique())-2, "Set1")
strict_arm_color <- c(strict_arm_color[1], "#E8E8E8", "#F5F5F5", strict_arm_color[-1])
names(strict_arm_color) <- anno_col$strict_arm %>% unique() %>% sort()

anno_color <- list(cancer_type = cancer_type_color,
                   target_pathway = target_pathway_color,
                   strict_arm  = strict_arm_color)

crxg_map_imputed %>%
  dplyr::select(-cosmic_id) %>%
  as.data.frame() %>%
  drop_na() %>%
  t() %>%
  pheatmap::pheatmap(na_col = "grey", scale = "column",
                     annotation_col = anno_col, 
                     annotation_row = anno_row,
                     annotation_colors = anno_color,
                     annotation_names_row = FALSE,
                     show_rownames = TRUE, 
                     show_colnames = FALSE,
                     cutree_rows = 7, 
                     filename = "crxg.pdf", 
                     width = 12, height = 6
                     )
                     
```

I export the allocation table 

```{r}
allocation %>% 
  mutate(value = 1) %>%
  spread(strict_arm, value, fill = 0) %>% 
  write_csv(here("data/allocation.csv"))
```

I create a ggridges figure 

```{r}
order <- crxg_map_imputed %>% 
  gather(drug, response, -cosmic_id) %>% 
  group_by(drug) %>% 
  summarise(sd = sd(response)) %>% 
  arrange(sd) %>% 
  .$drug

crxg_map_imputed %>% 
  gather(drug, response, -cosmic_id) %>% 
  mutate(drug = factor(drug, levels = order)) %>%
  ggplot(aes(y = drug, x = response, fill = ..x..)) + 
  ggridges::geom_density_ridges_gradient(scale = 2, rel_min_height = 0.01) +
  theme_ridges() + 
  scale_fill_gradient2() + 
  labs(x = "centered ln(IC50)",
       y = "Drug",
       fill = "Response Score",
       title = "Distribution of Response Scores by Drug",
       caption = "990 cell lines, median imputation") +
  ggsave("ridge_crxg.pdf", width = 6, height = 4)
```
