==> vignettes/6_rlearner.Rmd <==
---
title: "R Learner"
author: "Niklas Rindtorff"
date: "4/3/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(rlearner)
library(randomizr)
library(here)
```

```{r}
allocation <- read_csv(here("data/allocation.csv")) %>% mutate(cosmic_id = as.character(cosmic_id))
covariates <- read_csv(here("data/covariates.csv")) %>% mutate(cosmic_id = as.character(cosmic_id))

load(here("data/crxg_map_imputed.Rdata"))
load(here("data/mutation.Rdata"))
load(here("data/expression.Rdata"))
load(here("data/cnv.Rdata"))

X <- read_csv(here("data/X_umap.csv"))
w <- read_csv(here("data/X_umap.csv"))
y_complete <- read_csv(here("data/y_complete.csv"))

```

I source a set of functions, I will aggregate all of this in a package later. 

```{r, message=FALSE}
list.files(here("R"), full.names = TRUE) %>% lapply(source)
```

Similar to many causal inference problems, we have three datasets: unit covariates *X*, assignment *w* and responses *y*. 


The R-Learner is per default designed to evaluate the CATE of binary treatments. That means, a unit is either treated with a control of an intervention. In our case, we will define a central control arm, Cisplatin, which is commonly used in cases in which no further information about a malignant tumor is available (CUP Syndrome). 

For each treatment, we will train a two-step R-learner and estimate the CATE for the targeted agent vs. Cisplatin. 

Performing this type of star-schema analysis comes with a number of constraints. First, the model will be trained on a smaller number of cases and will likely be less generalizable to CATE estimation problems involving other therapies. Second, this way of analysis does not allow epsilon to be greater than 0, especially if we plan to supply the treatment propensities to the R-Learner.

We define three functions around the R-learner: 
1. A function that estimates the *w* parameter for weighted allocation
1. A function that will take a matrix of assignments based on therapeutic protocols and will introduce random re-assignments based on the parameters epsilon and gamma. 
1. A function that will take the assignment matrix and train R-learners for every treatment arm
1. A function that will collect the predicted CATEs for each treatment arm and calculate an aggregated performance metric

```{r}
estimate_cross(w, epsilon = 0, gamma = 1, omega = 0.1) # %>% knitr::kable()
```

```{r}
re_allocate(w, epsilon = 0, gamma = 0.5, omega = 0.5, ctrl = "cisplatin", return_propensity = FALSE) %>% 
  mutate(same = assignment == re_assignment)
```


```{r}
draw_interval = 0.05
n_draws = 5

x = rep(seq(0,1, draw_interval), each = n_draws)

allocation_dynamics_gamma <-  x %>% #gamma is scaling automatically
  parallel::mclapply(., re_allocate, allocation = w, ctrl= "cisplatin", epsilon = 0, omega = 0, return_propensity =FALSE) %>% 
  bind_rows() %>% 
  cbind(., gamma = rep(x, each = nrow(w))) %>%
  as_tibble() 

allocation_dynamics_gamma_count <- allocation_dynamics_gamma %>%
  count(gamma, re_assignment) %>% 
  mutate(n = n/n_draws)

allocation_dynamics_gamma_count %>% 
  ggplot(aes(gamma, n, color = re_assignment)) + 
  geom_point() + 
  geom_line() +
  theme_classic() + 
  scale_color_brewer(type = "qual") + 
   scale_y_log10(limits = c(1, 1000)) +
  labs(x = "Cross-In rate",
       y = "Number of units per treatment",
       title = "Treatment assignment by Cross-In rate",
       subtitle = "Cisplatin is the control group",
       caption = "Cross-In rate gamma, 20 draws",
       color = "Treatment") + 
  ggsave("gamma.pdf", width = 5, height = 4)
```

```{r}
draw_interval = 0.05
n_draws = 5

x = rep(seq(0,1, draw_interval), each = n_draws)

allocation_dynamics_epsilon <-  x %>% #gamma is scaling automatically
  parallel::mclapply(., re_allocate, allocation = w, ctrl= "cisplatin", gamma = 0, omega = 0,return_propensity =FALSE) %>% 
  bind_rows() %>% 
  cbind(., epsilon = rep(x, each = nrow(w))) %>%
  as_tibble() 

allocation_dynamics_epsilon_count <- allocation_dynamics_epsilon %>%
  count(omega, re_assignment) %>% 
  mutate(n = n/n_draws)

allocation_dynamics_epsilon_count %>% 
  ggplot(aes(epsilon, n, color = re_assignment)) + 
  geom_point() + 
  geom_line() +
  theme_classic() + 
  scale_color_brewer(type = "qual") + 
   scale_y_log10(limits = c(1, 1000)) +
  labs(x = "Cross-Over rate",
       y = "Number of units per treatment",
       title = "Treatment assignment by Cross-Over rate",
       subtitle = "Cisplatin is the control group",
       caption = "Cross-In rate epsilon, 20 draws",
       color = "Treatment") + 
  ggsave("epsilon.pdf", width = 5, height = 4)
```

```{r}
draw_interval = 0.05
n_draws = 5

x = rep(seq(0,1, draw_interval), each = n_draws)

allocation_dynamics_omega <-  x %>% #gamma is scaling automatically
  parallel::mclapply(., re_allocate, allocation = w, ctrl= "cisplatin", epsilon = 0, gamma = 0,return_propensity =FALSE) %>%
  bind_rows() %>% 
  cbind(., omega = rep(x, each = nrow(w))) %>%
  as_tibble() 

allocation_dynamics_omega_count <- allocation_dynamics_omega %>%
  count(omega, re_assignment) %>% 
  mutate(n = n/n_draws)

allocation_dynamics_omega_count %>% 
  ggplot(aes(omega, n, color = re_assignment)) + 
  geom_point() + 
  geom_line() +
  theme_classic() + 
  scale_color_brewer(type = "qual") + 
  scale_y_log10(limits = c(1, 1000)) +
  labs(x = "Cross-Out rate",
       y = "Number of units per treatment",
       title = "Treatment assignment by Cross-Out rate",
       subtitle = "Cisplatin is the control group",
       caption = "Cross-Out rate omega, 20 draws",
       color = "Treatment") + 
  ggsave("omega.pdf", width = 5, height = 4)
```


In case of a star-schema analysis, I need to introduce a third parameter, a cross-out rate omega. This parameter describes the rate at which a unit that matched a therapeutic protocol, after applying the cross_over rate,  is allocated into the control arm. Otherwise, these units had a treatment propensity for the control treatment of 0, thereby violating the positivity assumption.

# Create 2x2 tables of reassignments 

```{r}
# I focus on one representative treatment, Trametinib 
# I then test different configurations of gamma and omega
# I visualize the final assignment and the initial assignment as a 2x2 table 
gamma_count <- expand.grid(gamma = rep(seq(0.1,0.9, 0.2), each = 1),
            omega = rep(seq(0.1,0.9, 0.2), each = 1)) %>% 
  mutate(new = furrr::future_map2(gamma, omega, ~ re_allocate(allocation = w,
                                                              gamma = .x,
                                                              omega = .y,
                                                              ctrl= "cisplatin", 
                                                              epsilon = 0,
                                                              return_propensity =FALSE)))

gamma_count_df <- gamma_count %>% 
  unnest(new) %>% 
  filter(assignment %in% c("cisplatin", "trametinib"),
         re_assignment %in% c("cisplatin", "trametinib")) %>% 
  count(assignment, re_assignment, gamma, omega) %>% 
  arrange(gamma) %>% 
  nest(-gamma, -omega) %>% 
  mutate(data = map(data, ~ .x %>% mutate(percent = n/sum(n)*100)))

# gamma_count_mat <- gamma_count %>% 
#   mutate(matrix = map(data, ~ .x %>% complete(assignment, re_assignment, fill = list(n = 0)) %>% 
#   spread(re_assignment, n) %>% 
#   as.data.frame() %>% 
#   tibble::column_to_rownames("assignment") %>% 
#   as.matrix()))

gamma_count_df %>% 
  unnest() %>% 
  mutate(n_log = log10(n)) %>%
  ggplot(aes(assignment, re_assignment, fill = n_log)) + 
  geom_tile() + 
  geom_text(aes(label = round(percent, 0)), color = "white") + 
  facet_grid(gamma ~ omega) + 
  scale_fill_viridis_c() + 
  theme_minimal() + 
  labs(x = "Initial assignment",
       y = "Final assignment",
       title = "Treatment assignment by Cross-Out and Cross-In rate",
       subtitle = "Percent units in white, Cisplatin is the control group",
       caption = "Cross-Out rate omega and Cross-In rate gamma, 5 draws",
       fill = "Number of units\n(log-scale)") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggsave("heatmap_allocation.pdf", width = 7, height = 6)
```


After defining a function to perform weighted reassignments, I have to generate result vectors based on a given reassignment. The function name is *create_y*. 

Now I have a first training run of the R-Learner, with a 20% test holdout

```{r}
out <- re_allocate(w, epsilon = 0, gamma = 0.6, omega = 0.25, ctrl = "cisplatin") %>% 
  feed_rlearner(., X, y_complete, var = "afatinib", ctrl = "cisplatin") %>% 
  train_rlearner(., estimate_phat = FALSE)
```


# Measuring the PEHE for treatment effects.

Different from the initial proposal, I did not modify the PEHE function for a reference-free context. Instead, as described above, I decided to pick a reference treatment, cisplatin.
I showcase the *estimate_pehe* function below on the first training run.

```{r}
estimate_pehe(out) %>% 
  group_by(re_assignment, var) %>% 
  summarise(pehe_lasso_sd = sd(pehe_lasso),
            pehe_boost_sd = sd(pehe_boost),
            pehe_lasso = mean(pehe_lasso),
            pehe_boost = mean(pehe_boost))
```


```{r}
estimate_pehe(out) %>% 
  mutate(logical = factor(logical)) %>%
  ggplot(aes(tau, tau_hat_lasso, color = re_assignment )) + 
  geom_point()
```


# Iterate R-Learner 


```{r}
rlearner_grid <- expand.grid(gamma = rep(seq(0,1, 0.05), each = 1),
                                omega = rep(seq(0,1, 0.05), each = 1),
                                drug = w[,-1] %>% colnames()) %>% 
  filter(gamma != 0, gamma != 1) %>% 
  filter(omega != 0, omega != 1) %>%
  filter(drug != "cisplatin") %>% 
  #filter(!(drug %in% c("afatinib", "dabrafenib", "olaparib"))) %>%
  mutate(drug = as.character(drug)) %>%
  mutate(gamma_omega = paste0(gamma, "_", omega)) 

tmp <- rlearner_grid %>%
  mutate(data = map2(gamma_omega,drug, ~ iterate_rlearner(gamma_omega = .x, var_in = .y))) %>% 
  mutate(pehe = map(data, ~ estimate_pehe(.x)))
```

I collect data 

```{r}
rlearner_grid_result <- tibble(path = list.files(here("data"), pattern = "2019-04", full.names = TRUE),
       name = list.files(here("data"), pattern = "2019-04")) %>% 
  rowwise() %>%
  mutate(gamma = str_split(name, pattern = "_") %>% unlist() %>% .[1] %>% as.numeric(),
         omega = str_split(name, pattern = "_") %>% unlist() %>% .[2] %>% as.numeric(),
         drug = str_split(name, pattern = "_") %>% unlist() %>% .[3],
         time = str_split(name, pattern = "_") %>% unlist() %>% .[4] %>% substr(1, nchar(.)-4)) %>% 
  ungroup() %>%
  mutate(rlearner = map(path, ~ readRDS(.x)))

rlearner_grid_result_pehe <- rlearner_grid_result %>% 
  mutate(pehe = map(rlearner, ~ estimate_pehe(.x)),
         metrics = map(pehe, ~ .x %>% group_by(re_assignment, var) %>% 
                         summarise(pehe_lasso_sd = sd(pehe_lasso),
                         pehe_boost_sd = sd(pehe_boost),
                         pehe_lasso = mean(pehe_lasso),
                         pehe_boost = mean(pehe_boost)))) 
  

## I collect the same data for a UMAP-transformed X, including gene expression data and CNVs
rlearner_grid_result_umap <- tibble(path = list.files(here("data"), pattern = "2019-05", full.names = TRUE),
       name = list.files(here("data"), pattern = "2019-05")) %>% 
  rowwise() %>%
  mutate(gamma = str_split(name, pattern = "_") %>% unlist() %>% .[1] %>% as.numeric(),
         omega = str_split(name, pattern = "_") %>% unlist() %>% .[2] %>% as.numeric(),
         drug = str_split(name, pattern = "_") %>% unlist() %>% .[3],
         time = str_split(name, pattern = "_") %>% unlist() %>% .[4] %>% substr(1, nchar(.)-4)) %>% 
  ungroup() %>%
  mutate(rlearner = map(path, ~ readRDS(.x)))

rlearner_grid_result_pehe_umap <- rlearner_grid_result_umap %>% 
  mutate(pehe = map(rlearner, ~ estimate_pehe(.x)),
         metrics = map(pehe, ~ .x %>% group_by(var) %>% 
                         summarise(pehe_lasso_sd = sd(pehe_lasso),
                         pehe_boost_sd = sd(pehe_boost),
                         pehe_lasso = mean(pehe_lasso),
                         pehe_boost = mean(pehe_boost)))) 
```


```{r}
df <- rlearner_grid_result_pehe_umap %>% 
  unnest(metrics) %>%
  dplyr::select(pehe_lasso, pehe_boost, pehe_boost_sd, pehe_lasso_sd, drug, re_assignment, gamma, omega) %>%
  mutate(re_assignment = if_else(re_assignment == "cisplatin", "ctrl", "treatment")) %>% 
  gather(method, pehe, -(pehe_boost_sd:omega)) %>% 
  gather(method_sd, pehe_sd, -(drug:pehe))
  

df %>% 
  ggplot(aes(gamma, pehe, color = method)) + 
  geom_point() + 
  facet_grid(re_assignment ~ drug) + 
  scale_y_log10() + 
  theme_bw()
```



# Defining clinical regret 

What is the value of this CI? 

I calculate the baseline regret that would exist if cell lines would be allocated by their initial therpeutic protocol only. For the rest of this publication, I refer to the standard regret as the "baseline regret".

```{r}
standard_regret <- tibble(drug = rlearner_regret_umap$drug %>% unique()) %>% 
  mutate(standard_regret = map(drug, ~ return_standard_regret(.x))) %>% 
  unnest(standard_regret)

standard_regret
```

I also create a function to measure the protocol regret. 

```{r}
protocol_regret <- rlearner_grid_result_pehe_umap %>% 
  mutate(protocol_regret = map(rlearner, ~ .x %>% return_protocol_regret)) %>% 
  dplyr::select(name, protocol_regret) %>% 
  unnest(protocol_regret)

```

Depending on the CATE estimate, we can direct a given unit either into the ctrl of treatment arm. We then can estimate the difference in response 

```{r}
rlearner_regret_umap <- rlearner_grid_result_pehe_umap %>% 
  unnest(pehe) %>%
  mutate(tau_treatment = tau < 0, # the treatment was more effective
         tau_hat_lasso_treatment = tau_hat_lasso < 0, # the lasso model suggests that the treatment is more effective 
         tau_hat_boost_treatment = tau_hat_boost < 0) %>% 
  mutate(lasso_regret = ifelse(tau_treatment != tau_hat_lasso_treatment, tau, 0), # give regret if prediction is wrong
         boost_regret = ifelse(tau_treatment != tau_hat_boost_treatment, tau, 0),
         # Protocol regret is the regret caused by reassigning the patients into treatment arms, that are not part of the therapeutic protocol
         #protocol_regret = ifelse((logical == 0 & tau < 0) | (logical == 1 & tau > 0), tau, 0), 
         #reassignment_regret = ifelse(assignment != re_assignment & logical == 0 & tau < 0, tau, 0),
         #protocol_regret_minus_lasso = protocol_regret - lasso_regret,
         #protocol_regret_minus_boost = protocol_regret - boost_regret,
         #baseline_regret = ifelse((assignment == "cisplatin" & tau < 0) | (assignment != "cisplatin" & tau > 0), tau, 0)
         ) %>% # special case in which patient was reassigned
  group_by(gamma, omega, drug, time, name) %>% 
  summarise(lasso_regret = sum(abs(lasso_regret)),
            boost_regret = sum(abs(boost_regret)),
            #protocol_regret = sum(abs(protocol_regret)),
            #reassignment_regret = sum(abs(reassignment_regret)),
            #protocol_regret_minus_lasso = protocol_regret - lasso_regret,
            #protocol_regret_minus_boost = protocol_regret - boost_regret,
            #baseline_regret = sum(abs(baseline_regret))
            ) %>% 
  left_join(standard_regret, by = "drug") %>% 
  left_join(protocol_regret, by = "name") %>% 
  mutate(regret_reduction_lasso = (lasso_regret+protocol_regret)-standard_regret,
         regret_reduction_boost = (boost_regret+protocol_regret)-standard_regret)
  

```



```{r}
rlearner_regret_umap %>% 
  gather(type, regret, -gamma, -omega, -drug, -time, -name) %>%
  filter(type != "regret_reduction_lasso",
         type != "regret_reduction_boost",
         type != "standard_regret",
         type != "protocol_regret") %>%
  ggplot(aes(gamma, omega, fill = regret)) + 
  geom_tile() + 
  facet_grid(drug ~ type) + 
  scale_fill_viridis_c() + 
  theme_minimal() + 
  labs(x = "Cross-In rate",
       y = "Cross-Out rate",
       title = "Cummulative regret by Cross-Out and Cross-In rate",
       subtitle = "Two CATE estimation methods, Cisplatin is the control group",
       caption = "Cross-Out rate omega and Cross-In rate gamma, 20 draws",
       fill = "Cummulative Regret") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggsave("heatmap_regret_cate.pdf", width = 5.5, height =10)
```

```{r}
rlearner_regret_umap %>% 
  gather(type, regret, -gamma, -omega, -drug, -time, -name) %>%
  filter(type == "protocol_regret") %>%
  ggplot(aes(gamma, omega, fill = regret)) + 
  geom_tile() + 
  facet_grid(drug ~ type) + 
  scale_fill_viridis_c() +
  theme_minimal() + 
  labs(x = "Cross-In rate",
       y = "Cross-Out rate",
       title = "Protocol regret by Cross-Out and Cross-In rate",
       subtitle = "Cisplatin is the control group",
       caption = "Cross-Out rate omega and Cross-In rate gamma, 20 draws",
       fill = "Cummulative Regret") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggsave("heatmap_regret_protocol.pdf", width = 4.0 , height =10)
```


```{r}
rlearner_regret_umap %>% 
  gather(type, regret, -gamma, -omega, -drug, -time, -name) %>%
  filter(type == "regret_reduction_lasso" |
         type == "regret_reduction_boost") %>%
  ggplot(aes(gamma, omega, fill = regret)) + 
  geom_tile() + 
  facet_grid(drug ~ type) + 
  scale_fill_gradient2() + 
  theme_minimal() + 
  labs(x = "Cross-In rate",
       y = "Cross-Out rate",
       title = "Overall reduction of regret by Cross-Out and Cross-In rate",
       subtitle = "Two CATE estimation methods, Cisplatin is the control group",
       caption = "Cross-Out rate omega and Cross-In rate gamma, 20 draws",
       fill = "Regret Reduction\ncompared to current standard") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggsave("heatmap_regret_reduction.pdf", width = 6 , height =10)
```

```{r}
rlearner_grid_result_pehe_umap %>% 
  
  unnest(metrics) %>% 
  dplyr::select(gamma, omega, drug,  pehe_lasso, pehe_boost) %>%
  gather(type, pehe, -gamma, -omega, -drug) %>%
  mutate(log_pehe = log10(pehe)) %>%
  ggplot(aes(gamma, omega, fill = log_pehe)) + 
  geom_tile() + 
  facet_grid(drug ~ type) + 
  scale_fill_viridis_c() + 
  theme_minimal() + 
  labs(x = "Cross-In rate",
       y = "Cross-Out rate",
       title = "Average PEHE by Cross-Out and Cross-In rate",
       subtitle = "Two CATE estimation methods, Cisplatin is the control group",
       caption = "Cross-Out rate omega and Cross-In rate gamma, 20 draws",
       fill = "Average PEHE\n(log-scale)") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggsave("heatmap_pehe.pdf", width = 5.2 , height =10)
  
```


```{r}
rlearner_grid_result_pehe_umap %>%
  unnest(pehe) %>% 
  group_by(gamma, omega, drug, re_assignment) %>% 
  summarise(var_lasso = var(tau_hat_lasso),
            var_boost = var(tau_hat_boost))  %>%
  dplyr::select(-re_assignment) %>%
  gather(type, var, -drug, -omega, -gamma) %>% 
  mutate(var = sqrt(var)) %>%
  ggplot(aes(gamma, omega, fill = var)) + 
  geom_tile() + 
  facet_grid(drug ~ type) + 
  scale_fill_viridis_c() + 
  theme_minimal() + 
  labs(x = "Cross-In rate",
       y = "Cross-Out rate",
       title = "Average Variance of Tau-estimate by Cross-Out and Cross-In rate",
       subtitle = "Two CATE estimation methods, Cisplatin is the control group",
       caption = "Cross-Out rate omega and Cross-In rate gamma, 20 draws",
       fill = "Variance of Tau\n(sqrt)") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  ggsave("heatmap_var_tau.pdf", width = 5.2 , height =10)
```



```{r}
rlearner_grid$pehe[[1]] %>% 
  ggplot(aes(tau_hat_lasso, tau, color = re_assignment)) + 
  geom_point() + 
  theme_classic()
```



```{r}
y_complete %>% 
  ggplot(aes(trametinib, cisplatin)) + 
  geom_point() + 
  theme_classic() + 
  labs(title = "Association of Cisplatin and Trametinib response")
```
